lsp-devtools: Rewrite the `record` command to use `LSPAgentClient`
This uses a modifed `RichHandler` from `rich` to output a live log to
the console, drawing data via the new `LSPAgentClient`.

I'm not convinced how useful the live console output is since a single
completion response from `esbonio` is enough to swamp the scrollback
buffer in the terminal. However, it at least proves the concept.

Eventually this command can be extended with options to record a
session to a file or SQLite database which I expect will be a lot more
useful!