lsp-devtools: Refactor agent
The agent itself has been greatly simplified!

Rather than run multiple pygls servers with multiple event loops in
multiple threads we now have a single asyncio loop that runs multiple
concurrent tasks.

- A task feeding the process' stdin to server stdin
- A task feeding server stdout to the process' stdout
- A task polling the server process to ensure its still alive

As an extra benefit, shutdown when there is a fatal exception is now
much cleaner. However, automatic shutdown after the client process
ends still needs implementing!

Additionally, interactions with the agent are starting to be
formalised with the introductions of the `AgentServer` and
`AgentClient`.

The `AgentServer` is run within the same process as the agent and will
provide the means for communication with other processes.

The `AgentClient` is run within other processes trying to interact
with the agent.