lsp-devtools: Swap relationship between client and server
Originally, the assumption was that the lsp agent would host a server
that commands like `lsp-devtools record` and `lsp-devtools tui`
would connect to.

However this posed a number of issues, such as commands like `record`
missing the beginning of a session since the agent would not wait for
a connection (#29) as well as the agent not stopping once an LSP
session ended (#17)

So this commit reverses that relationship, with commands like
``record`` spinning up the server and the lsp agent creating a client
connection.

Additionally this commit
- Switches to using the base `Client` provided by a future version of
  `pygls`
- Trials using async stdin/stdout streams for the agent - a potential
  candidate someday for upstreaming into `pygls`
- Switches to using a TCP connection between client and server. (#37)