lsp-devtools: Proof of concept language client
This adds a proof of concept language client built on textual's
TextArea. The idea is to eventually be able to "play" with a server
while inpersonating a particular client - all while being able to
inspect the traffic sent to/from the server.

The low-level stuff is technically done? The server starts, we can
capture and display the traffic - even text sync appears to work!

It's now "just" a case of implementing a UI and adding support for
all the LSP methods!