docs: Add support for client capabilities with values
This adds a `values-table` directive that can generate tables
summarizing a set of values e.g. `SymbolKind` that is supported by the
client.

This commit also refactors the `capabailities` extension in general to
use the `ObjectDescription` directive, which aside from looking nicer,
also allows us to easily generate toc entries for each capability.