pytest-lsp: Introduce `register_lsp_features`
To make the process of customising the language client easier the
`register_lsp_features` function has been introduced, which accepts a
client instance and a map from method name to the corresponding
implementation.

Coupled with recent fixes in pygls, it should now be possible to
override arbitrary methods and use a custom class instance.